refactor(messages): extract shared polling hook + layout to dedupe containers
購入者向け messages-container.tsx と販売者向け seller-messages-container.tsx の
~214 行の相互コピーを解消する（SonarCloud Duplicated Lines 9.7% → ゲート Failed の主因）。

- src/components/shared/messages/use-conversation-thread.ts: ポーリング/既読化/
  送信後再フェッチ/レースガードを集約。ログ出所はラベル引数で区別し既存ログ文言を維持。
- src/components/shared/messages/messages-layout.tsx: 2ペイン骨格を汎用化。唯一異なる
  会話行のアバター取得元を getAvatar アダプタで注入（購入者=店舗 / 販売者=購入者）。
- 両コンテナを薄いラッパに置換（public props・default export 名は不変、呼び出し側無改修）。
- 切替時クリア挙動とレースガードは購入者側の優れた実装に統一。

挙動不変の refactor。既存21テスト（messages-container/seller/conversation-thread）緑、
tsc/lint クリーン。props は S6759 対応で Readonly 化。
